+++ src/App.tsx
@@ -1,10 +1,11 @@
-import React, { useEffect, useRef } from "react";
+import React from "react";
 import "./App.css";
 //import { Garden } from "./interfaces/garden";
 import { Plant } from "./interfaces/plant";
 import { PlantView } from "./PlantView";
 import { shadeLevel } from "./interfaces/shadeLevel";
 import Serviceberry from "./images/Amelanchier_x_grandifolia_Autumn_Brilliance_1.jpg";
+import Garden from "./components/Garden";
 
 function App(): JSX.Element {
     /* So, in Quizzer, there is an initial list of quizzes given by "sample",
@@ -35,67 +36,6 @@ function App(): JSX.Element {
     //const [plants, setPlants] = useState<Plant[]>([firstPlant]);
 
     // Lines 35-95 is the DnD (Drag and Drop) implementation for our application
-    const containerRef = useRef<HTMLDivElement>(null);
-    const PropRef = useRef<HTMLDivElement>(null);
-
-    const isClicked = useRef<boolean>(false);
-
-    const coords = useRef<{
-        startX: number;
-        startY: number;
-        lastX: number;
-        lastY: number;
-    }>({
-        startX: 0,
-        startY: 0,
-        lastX: 0,
-        lastY: 0
-    });
-
-    useEffect(() => {
-        if (!PropRef.current || !containerRef.current) return;
-
-        const prop = PropRef.current;
-        const container = containerRef.current;
-
-        const onMouseDown = (e: MouseEvent) => {
-            isClicked.current = true;
-            coords.current.startX = e.clientX;
-            coords.current.startY = e.clientY;
-        };
-        // I deleted "e: MouseEvent" in the next function definition because we do not reference "e".
-        const onMouseUp = () => {
-            isClicked.current = false;
-            coords.current.lastX = prop.offsetLeft;
-            coords.current.lastY = prop.offsetTop;
-        };
-
-        const onMouseMove = (e: MouseEvent) => {
-            if (!isClicked.current) return;
-
-            const nextX =
-                e.clientX - coords.current.startX + coords.current.lastX;
-            const nextY =
-                e.clientY - coords.current.startY + coords.current.lastY;
-
-            prop.style.top = `${nextY}px`;
-            prop.style.left = `${nextX}px`;
-        };
-
-        prop.addEventListener("mousedown", onMouseDown);
-        prop.addEventListener("mouseup", onMouseUp);
-        container.addEventListener("mousemove", onMouseMove);
-        container.addEventListener("mouseleave", onMouseUp);
-
-        const cleanup = () => {
-            prop.removeEventListener("mousedown", onMouseDown);
-            prop.removeEventListener("mouseup", onMouseUp);
-            container.removeEventListener("mousemove", onMouseMove);
-            container.removeEventListener("mouseleave", onMouseUp);
-        };
-
-        return cleanup;
-    }, []);
 
     return (
         <div className="App">
@@ -107,6 +47,7 @@ function App(): JSX.Element {
                 automatically reload.
             </p>
             <PlantView plant={firstPlant}></PlantView>
+            <Garden></Garden>
         </div>
     );
 }
