+++ src/components/Prop.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Prop() {
+    return <div>Prop</div>;
+}
+
+export default Prop;
